# specs/api.yml
openapi: 3.0.0
info:
  title: OnePage 電商平台 API
  version: 1.0.0
  description: |
    OnePage 一頁式電商平台 API 契約。
    後端：Spring Boot 3.2 + JPA + PostgreSQL。
    金流：綠界全方位金流 API v5。
    發票：綠界電子發票 API。
servers:
  - url: /api/v1
    description: API v1
...
paths:
...
  /product-categories:
    get:
      tags: [ProductCategory]
      summary: 查詢商品類型清單（樹狀結構）
      description: 管理員返回所有類型；一般使用者僅返回自己建立的類型（REQ-034 RBAC 資料隔離）
      operationId: listProductCategories
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 分類樹（依 RBAC 角色過濾）
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductCategoryDTO'

    post:
      tags: [ProductCategory]
      summary: 新增商品類型
      operationId: createProductCategory
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductCategoryRequest'
      responses:
        '201':
          description: 建立成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductCategoryDTO'
        '400':
          description: 驗證失敗
  /product-categories/{categoryId}:
    put:
      tags: [ProductCategory]
      summary: 修改商品類型名稱（不可調整父子關係）
      operationId: updateProductCategory
      security:
        - bearerAuth: []
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 100
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductCategoryDTO'
        '400':
          description: 驗證失敗

    delete:
      tags: [ProductCategory]
      summary: 刪除商品類型
      operationId: deleteProductCategory
      security:
        - bearerAuth: []
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: 刪除成功
        '400':
          description: 該分類下仍有商品，無法刪除

  # ═══ Product ═══
  /products:
    get:
      tags: [Product]
      summary: 查詢商品清單（後台）
      description: 管理員返回所有商品；一般使用者僅返回自己建立的商品（REQ-034 RBAC 資料隔離）
      operationId: listProducts
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [ACTIVE, INACTIVE]
        - name: categoryId
          in: query
          schema:
            type: integer
            format: int64
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 商品清單（依 RBAC 角色過濾）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedProducts'

    post:
      tags: [Product]
      summary: 上架商品
      operationId: createProduct
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CreateProductRequest'
      responses:
        '201':
          description: 商品上架成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDTO'
        '400':
          description: 驗證失敗
  /products/{productId}:
    get:
      tags: [Product]
      summary: 查詢單一商品（後台）
      operationId: getProduct
      security:
        - bearerAuth: []
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 商品詳情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDTO'
        '404':
          description: 商品不存在

    put:
      tags: [Product]
      summary: 編輯商品資訊（不含庫存）
      operationId: updateProduct
      security:
        - bearerAuth: []
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UpdateProductRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDTO'
        '400':
          description: 驗證失敗
        '404':
          description: 商品不存在
  /products/{productId}/images:
    post:
      tags: [Product]
      summary: 上傳商品圖片（REQ-026）
      operationId: uploadProductImages
      security:
        - bearerAuth: []
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [images]
              properties:
                images:
                  type: array
                  items:
                    type: string
                    format: binary
                  maxItems: 5
      responses:
        '200':
          description: 圖片上傳成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductImageDTO'
        '400':
          description: 圖片數量超過上限或格式不符
        '404':
          description: 商品不存在
  /products/{productId}/deactivate:
    post:
      tags: [Product]
      summary: 下架商品
      operationId: deactivateProduct
      security:
        - bearerAuth: []
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 下架成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDTO'
        '400':
          description: 僅已上架商品可被下架
        '404':
          description: 商品不存在

  # ═══ Website ═══
...
  /websites/{websiteId}:
    get:
      tags: [Website]
      summary: 查詢單一網站
      operationId: getWebsite
      security:
        - bearerAuth: []
      parameters:
        - name: websiteId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 網站詳情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebsiteDTO'
        '404':
          description: 網站不存在

    put:
      tags: [Website]
      summary: 編輯網站
      operationId: updateWebsite
      security:
        - bearerAuth: []
      parameters:
        - name: websiteId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UpdateWebsiteRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebsiteDTO'
        '400':
          description: 驗證失敗
...
  /storefront/websites/{websiteId}/products/{productSlug}:
    get:
      tags: [Storefront]
      summary: 取得商品詳情頁資料
      operationId: getStorefrontProduct
      parameters:
        - name: websiteId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: productSlug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 商品詳情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StorefrontProductDTO'
        '404':
          description: 商品不存在或未上架至此網站
...
  /storefront/orders:
    post:
      tags: [Order]
      summary: 建立訂單（前台顧客）
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: 訂單建立成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDTO'
        '400':
          description: 驗證失敗或庫存不足
        '409':
          description: 庫存衝突（樂觀鎖）
  /orders:
    get:
      tags: [Order]
      summary: 查詢訂單清單（後台）
      description: 管理員返回所有訂單；一般使用者僅返回自己網站的訂單（REQ-034 RBAC 資料隔離）
      operationId: listOrders
      security:
        - bearerAuth: []
      parameters:
        - name: websiteId
          in: query
          schema:
            type: integer
            format: int64
        - name: orderNumber
          in: query
          schema:
            type: string
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING_PAYMENT, PROCESSING_PAYMENT, PAID, SHIPPED, RETURNED, PAYMENT_FAILED]
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 訂單清單
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedOrders'
  /orders/export:
    get:
      tags: [Order]
      summary: 匯出訂單 CSV
      operationId: exportOrdersCsv
      security:
        - bearerAuth: []
      parameters:
        - name: websiteId
          in: query
          schema:
            type: integer
            format: int64
        - name: orderNumber
          in: query
          schema:
            type: string
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING_PAYMENT, PROCESSING_PAYMENT, PAID, SHIPPED, RETURNED, PAYMENT_FAILED]
      responses:
        '200':
          description: CSV 檔案（UTF-8 with BOM）
          content:
            text/csv:
              schema:
                type: string
                format: binary
  /orders/{orderId}/ship:
    post:
      tags: [Order]
      summary: 標註出貨
      operationId: markOrderShipped
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 標註出貨成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDTO'
        '400':
          description: 僅已付款訂單可標註出貨
  /orders/{orderId}/return:
    post:
      tags: [Order]
      summary: 標註退貨
      operationId: markOrderReturned
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 標註退貨成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDTO'
        '400':
          description: 僅已付款訂單可標註退貨

  # ═══ Payment ═══
  /storefront/orders/{orderId}/payment:
    post:
      tags: [Payment]
      summary: 建立付款（選擇付款方式 + 發票資訊）
      operationId: createPayment
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePaymentRequest'
      responses:
        '201':
          description: 付款建立成功，回傳綠界付款頁面 URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentDTO'
        '400':
          description: 僅待付款訂單可建立付款
  /payment/ecpay/callback:
    post:
      tags: [Payment]
      summary: 綠界付款回調（由綠界呼叫）
      operationId: ecpayCallback
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              description: 綠界回調參數（含 CheckMacValue）
      responses:
        '200':
          description: 回應 "1|OK"
          content:
            text/plain:
              schema:
                type: string
                example: "1|OK"

  # ═══ Invoice ═══
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ProductCategoryDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        parentId:
          type: integer
          format: int64
          nullable: true
        ownerUserId:
          type: integer
          format: int64
          description: 建立者 user ID（REQ-034）
        children:
          type: array
          items:
            $ref: '#/components/schemas/ProductCategoryDTO'
    CreateProductCategoryRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        parentId:
          type: integer
          format: int64
          nullable: true
          description: 父分類 ID（null 表示新增主分類）

    # ── Product ──
    ProductDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        ownerUserId:
          type: integer
          format: int64
          description: 建立者 user ID（REQ-034）
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        price:
          type: number
          format: double
        priceUnit:
          type: string
          enum: [KG, CATTY]
        packaging:
          type: string
        categoryId:
          type: integer
          format: int64
        categoryName:
          type: string
        status:
          type: string
          enum: [ACTIVE, INACTIVE]
        isBundle:
          type: boolean
        bundleDiscountPercent:
          type: number
          format: double
          nullable: true
        bundleItems:
          type: array
          items:
            $ref: '#/components/schemas/BundleItemDTO'
        isPreorder:
          type: boolean
        preorderStartDate:
          type: string
          format: date
          nullable: true
        preorderEndDate:
          type: string
          format: date
          nullable: true
        preorderDiscountPercent:
          type: number
          format: double
          nullable: true
        shippingDeadline:
          type: string
          format: date
          nullable: true
        stockQuantity:
          type: integer
        images:
          type: array
          items:
            $ref: '#/components/schemas/ProductImageDTO'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    BundleItemDTO:
      type: object
      properties:
        productId:
          type: integer
          format: int64
        productName:
          type: string
        price:
          type: number
          format: double
    ProductImageDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        imageUrl:
          type: string
        sortOrder:
          type: integer
    CreateProductRequest:
      type: object
      required: [name, price, priceUnit, categoryId, stockQuantity]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        price:
          type: number
          format: double
          minimum: 0.01
        priceUnit:
          type: string
          enum: [KG, CATTY]
        packaging:
          type: string
        categoryId:
          type: integer
          format: int64
        stockQuantity:
          type: integer
          minimum: 0
        images:
          type: array
          items:
            type: string
            format: binary
          maxItems: 5
        isBundle:
          type: boolean
          default: false
        bundleDiscountPercent:
          type: number
          format: double
        bundleProductIds:
          type: array
          items:
            type: integer
            format: int64
        isPreorder:
          type: boolean
          default: false
        preorderStartDate:
          type: string
          format: date
        preorderEndDate:
          type: string
          format: date
        preorderDiscountPercent:
          type: number
          format: double
    UpdateProductRequest:
      type: object
      required: [name, price, priceUnit, categoryId]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        price:
          type: number
          format: double
          minimum: 0.01
        priceUnit:
          type: string
          enum: [KG, CATTY]
        packaging:
          type: string
        categoryId:
          type: integer
          format: int64
        images:
          type: array
          items:
            type: string
            format: binary
          maxItems: 5
        isBundle:
          type: boolean
        bundleDiscountPercent:
          type: number
          format: double
        bundleProductIds:
          type: array
          items:
            type: integer
            format: int64
        isPreorder:
          type: boolean
        preorderStartDate:
          type: string
          format: date
        preorderEndDate:
          type: string
          format: date
        preorderDiscountPercent:
          type: number
          format: double
        shippingDeadline:
          type: string
          format: date
    PagedProducts:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/ProductDTO'
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        page:
          type: integer
        size:
          type: integer

    # ── Website ──
    WebsiteDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        title:
          type: string
          description: 網站標題（前台顯示用）（REQ-029）
        subtitle:
          type: string
          nullable: true
          description: 網站副標題（選填）（REQ-029）
        browserTitle:
          type: string
          nullable: true
          description: 瀏覽器分頁標題 <title> tag（選填）（REQ-029）
        footerTitle:
          type: string
          nullable: true
          description: Footer 標題（店名）（選填）（REQ-029）
        footerSubtitle:
          type: string
          nullable: true
          description: Footer 副標題（聯絡資訊）（選填）（REQ-029）
        subscriptionPlan:
          type: string
        publishStartAt:
          type: string
          format: date-time
          nullable: true
        publishEndAt:
          type: string
          format: date-time
          nullable: true
        bannerImageUrl:
          type: string
          nullable: true
        promoImageUrl:
          type: string
          nullable: true
        freeShippingThreshold:
          type: number
          format: double
        status:
          type: string
          enum: [DRAFT, PUBLISHED, OFFLINE]
        storefrontUrl:
          type: string
          description: 前台 URL /project/{websiteId}
        ownerUserId:
          type: integer
          format: int64
          description: 建立者 user ID（RBAC 資料隔離用）（REQ-034）
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    UpdateWebsiteRequest:
      type: object
      required: [name, title]
      properties:
        name:
          type: string
          minLength: 1
        title:
          type: string
          minLength: 1
          description: 網站標題（必填）（REQ-029）
        subtitle:
          type: string
          nullable: true
          description: 網站副標題（選填）（REQ-029）
        browserTitle:
          type: string
          nullable: true
          description: 瀏覽器分頁標題（選填）（REQ-029）
        footerTitle:
          type: string
          nullable: true
          description: Footer 標題（選填）（REQ-029）
        footerSubtitle:
          type: string
          nullable: true
          description: Footer 副標題（選填）（REQ-029）
        subscriptionPlan:
          type: string
        publishStartAt:
          type: string
          format: date-time
        publishEndAt:
          type: string
          format: date-time
        bannerImage:
          type: string
          format: binary
          nullable: true
          description: Banner 主視覺圖片（REQ-029）
        promoImage:
          type: string
          format: binary
          nullable: true
          description: 宣傳文案圖片（選填）（REQ-029）
        freeShippingThreshold:
          type: number
          format: double
...
    StorefrontProductDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        price:
          type: number
          format: double
        priceUnit:
          type: string
          enum: [KG, CATTY]
        packaging:
          type: string
        images:
          type: array
          items:
            $ref: '#/components/schemas/ProductImageDTO'
        isPreorder:
          type: boolean
        preorderStartDate:
          type: string
          format: date
          nullable: true
        preorderEndDate:
          type: string
          format: date
          nullable: true
        preorderDiscountPercent:
          type: number
          format: double
          nullable: true
        isBundle:
          type: boolean
        bundleDiscountPercent:
          type: number
          format: double
          nullable: true
        bundleItems:
          type: array
          items:
            $ref: '#/components/schemas/BundleItemDTO'
        websiteId:
          type: integer
          format: int64

    # ── Shipping ──
...
    CreateOrderRequest:
      type: object
      required: [websiteId, customerName, customerPhone, customerEmail, shippingMethod, items]
      properties:
        websiteId:
          type: integer
          format: int64
        customerName:
          type: string
          minLength: 1
        customerPhone:
          type: string
          minLength: 1
        customerEmail:
          type: string
          format: email
        shippingAddress:
          type: string
          description: 運送地址（宅配必填，自取可空）
        shippingMethod:
          type: string
          enum: [DELIVERY, PICKUP]
        note:
          type: string
        taxId:
          type: string
          pattern: '^\d{8}$'
          description: 統一編號（選填，8碼數字）
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/OrderItemInput'
    OrderItemInput:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: integer
          format: int64
        quantity:
          type: integer
          minimum: 1
    OrderDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        orderNumber:
          type: string
        websiteId:
          type: integer
          format: int64
        websiteName:
          type: string
        customerName:
          type: string
        customerPhone:
          type: string
        customerEmail:
          type: string
        shippingAddress:
          type: string
        shippingMethod:
          type: string
          enum: [DELIVERY, PICKUP]
        shippingFee:
          type: number
          format: double
        subtotal:
          type: number
          format: double
        totalAmount:
          type: number
          format: double
        note:
          type: string
        taxId:
          type: string
        status:
          type: string
          enum: [PENDING_PAYMENT, PROCESSING_PAYMENT, PAID, SHIPPED, RETURNED, PAYMENT_FAILED]
        isPreorder:
          type: boolean
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemDTO'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    PagedOrders:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/OrderDTO'
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        page:
          type: integer
        size:
          type: integer

    # ── Payment ──
...